requirements-prod.txt
fastapi
uvicorn[standard]
gunicorn
pydantic-settings
curl_cffi
SQLAlchemy
alembic
psycopg[binary]
pandas
numpy
scikit-learn==1.7.2
joblib
scipy
xgboost
lightgbm
lxml
torch
matplotlib
python-dotenv
modal
